Refactor to `App\Controller\ForkController.php` and `App\Component\Application\ApplicationInterface.php`
