+++ src/Backend/Core/Engine/Ajax.php
@@ -2,7 +2,7 @@
 
 namespace Backend\Core\Engine;
 
-use ForkCMS\App\ApplicationInterface;
+use App\Component\Application\ApplicationInterface;
 use ForkCMS\App\KernelLoader;
 use Symfony\Component\HttpFoundation\Response;
 use Backend\Core\Engine\Base\AjaxAction as BackendBaseAJAXAction;
